+++ lua/lazyvim/config/init.lua
@@ -3,7 +3,7 @@ local Util = require("lazyvim.util")
 ---@class LazyVimConfig: LazyVimOptions
 local M = {}
 
-M.version = "10.2.0" -- x-release-please-version
+M.version = "10.5.0" -- x-release-please-version
 
 ---@class LazyVimOptions
 local defaults = {
@@ -92,7 +92,7 @@ local defaults = {
       Variable      = "󰀫 ",
     },
   },
-  ---@type table<string, string[]>?
+  ---@type table<string, string[]|boolean>?
   kind_filter = {
     default = {
       "Class",
@@ -109,6 +109,8 @@ local defaults = {
       "Struct",
       "Trait",
     },
+    markdown = false,
+    help = false,
     -- you can specify a different filter for each filetype
     lua = {
       "Class",
@@ -178,10 +180,7 @@ function M.setup(opts)
 
       Util.format.setup()
       Util.news.setup()
-
-      vim.api.nvim_create_user_command("LazyRoot", function()
-        Util.root.info()
-      end, { desc = "LazyVim roots for the current buffer" })
+      Util.root.setup()
 
       vim.api.nvim_create_user_command("LazyExtras", function()
         Util.extras.show()
@@ -214,7 +213,11 @@ function M.get_kind_filter(buf)
   if M.kind_filter == false then
     return
   end
-  return M.kind_filter[ft] or M.kind_filter.default
+  if M.kind_filter[ft] == false then
+    return
+  end
+  ---@diagnostic disable-next-line: return-type-mismatch
+  return type(M.kind_filter) == "table" and type(M.kind_filter.default) == "table" and M.kind_filter.default or nil
 end
 
 ---@param name "autocmds" | "options" | "keymaps"