+++ lua/lazyvim/plugins/extras/coding/tabnine.lua
@@ -5,10 +5,12 @@ return {
   {
     "nvim-cmp",
     dependencies = {
-      -- Add TabNine support, make sure you run :CmpTabnineHub after installation.
       {
         "tzachar/cmp-tabnine",
-        build = "./install.sh",
+        build = {
+          Util.is_win() and "pwsh -noni .\\install.ps1" or "./install.sh",
+          ":CmpTabnineHub",
+        },
         dependencies = "hrsh7th/nvim-cmp",
         opts = {
           max_lines = 1000,
@@ -16,8 +18,7 @@ return {
           sort = true,
         },
         config = function(_, opts)
-          local tabnine = require("cmp_tabnine.config")
-          tabnine:setup(opts)
+          require("cmp_tabnine.config"):setup(opts)
         end,
       },
     },