+++ lua/lazyvim/plugins/extras/editor/symbols-outline.lua
@@ -12,14 +12,20 @@ return {
         symbols = {},
         symbol_blacklist = {},
       }
+      local filter = Config.kind_filter
 
-      for kind, symbol in pairs(defaults.symbols) do
-        opts.symbols[kind] = {
-          icon = Config.icons.kinds[kind] or symbol.icon,
-          hl = symbol.hl,
-        }
-        if not vim.tbl_contains(Config.kind_filter.default, kind) then
-          table.insert(opts.symbol_blacklist, kind)
+      if type(filter) == "table" then
+        filter = filter.default
+        if type(filter) == "table" then
+          for kind, symbol in pairs(defaults.symbols) do
+            opts.symbols[kind] = {
+              icon = Config.icons.kinds[kind] or symbol.icon,
+              hl = symbol.hl,
+            }
+            if not vim.tbl_contains(filter, kind) then
+              table.insert(opts.symbol_blacklist, kind)
+            end
+          end
         end
       end
       return opts