+++ lua/lazyvim/plugins/extras/lang/markdown.lua
@@ -11,7 +11,7 @@ return {
     "williamboman/mason.nvim",
     opts = function(_, opts)
       opts.ensure_installed = opts.ensure_installed or {}
-      vim.list_extend(opts.ensure_installed, { "markdownlint" })
+      vim.list_extend(opts.ensure_installed, { "markdownlint", "marksman" })
     end,
   },
   {
@@ -54,11 +54,26 @@ return {
         "<leader>cp",
         ft = "markdown",
         "<cmd>MarkdownPreviewToggle<cr>",
-        desc = "Peek (Markdown Preview)",
+        desc = "Markdown Preview",
       },
     },
     config = function()
       vim.cmd([[do FileType]])
     end,
+  },
+
+  {
+    "lukas-reineke/headlines.nvim",
+    opts = function()
+      local opts = {}
+      for _, ft in ipairs({ "markdown", "norg", "rmd", "org" }) do
+        opts[ft] = { headline_highlights = {} }
+        for i = 1, 6 do
+          table.insert(opts[ft].headline_highlights, "Headline" .. i)
+        end
+      end
+      return opts
+    end,
+    ft = { "markdown", "norg", "rmd", "org" },
   },
 }