+++ lua/lazyvim/plugins/extras/lang/typescript.lua
@@ -32,6 +32,19 @@ return {
               end,
               desc = "Organize Imports",
             },
+            {
+              "<leader>cR",
+              function()
+                vim.lsp.buf.code_action({
+                  apply = true,
+                  context = {
+                    only = { "source.removeUnused.ts" },
+                    diagnostics = {},
+                  },
+                })
+              end,
+              desc = "Remove Unused Imports",
+            },
           },
           settings = {
             typescript = {