+++ lua/lazyvim/plugins/extras/ui/alpha.lua
@@ -27,9 +27,9 @@ return {
         dashboard.button("n", " " .. " New file",        "<cmd> ene <BAR> startinsert <cr>"),
         dashboard.button("r", " " .. " Recent files",    "<cmd> Telescope oldfiles <cr>"),
         dashboard.button("g", " " .. " Find text",       "<cmd> Telescope live_grep <cr>"),
-        dashboard.button("c", " " .. " Config",          "<cmd> e $MYVIMRC <cr>"),
+        dashboard.button("c", " " .. " Config",          "<cmd> lua require('lazyvim.util').telescope.config_files()() <cr>"),
         dashboard.button("s", " " .. " Restore Session", [[<cmd> lua require("persistence").load() <cr>]]),
-        dashboard.button("e", " " .. " Lazy Extras",     "<cmd> LazyExtras <cr>"),
+        dashboard.button("x", " " .. " Lazy Extras",     "<cmd> LazyExtras <cr>"),
         dashboard.button("l", "󰒲 " .. " Lazy",            "<cmd> Lazy <cr>"),
         dashboard.button("q", " " .. " Quit",            "<cmd> qa <cr>"),
       }