+++ lua/lazyvim/plugins/extras/util/dot.lua
@@ -37,7 +37,7 @@ return {
     "nvim-treesitter/nvim-treesitter",
     opts = function(_, opts)
       local function add(lang)
-        if type(opts.ensure_installed) ~= "table" then
+        if type(opts.ensure_installed) == "table" then
           table.insert(opts.ensure_installed, lang)
         end
       end