+++ lua/lazyvim/plugins/formatting.lua
@@ -2,22 +2,15 @@ local Util = require("lazyvim.util")
 
 local M = {}
 
----@type ConformOpts
-local format_opts = {}
-
 ---@param opts ConformOpts
 function M.setup(_, opts)
-  local util = require("conform.util")
-  opts.formatters = opts.formatters or {}
-  for name, formatter in pairs(opts.formatters) do
+  for name, formatter in pairs(opts.formatters or {}) do
     if type(formatter) == "table" then
-      local ok, defaults = pcall(require, "conform.formatters." .. name)
-      if ok and type(defaults) == "table" then
-        opts.formatters[name] = vim.tbl_deep_extend("force", {}, defaults, formatter)
-      end
-      if opts.formatters[name].extra_args then
-        opts.formatters[name].args =
-          util.extend_args(opts.formatters[name].args or {}, opts.formatters[name].extra_args)
+      ---@diagnostic disable-next-line: undefined-field
+      if formatter.extra_args then
+        ---@diagnostic disable-next-line: undefined-field
+        formatter.prepend_args = formatter.extra_args
+        Util.deprecate(("opts.formatters.%s.extra_args"):format(name), ("opts.formatters.%s.prepend_args"):format(name))
       end
     end
   end
@@ -29,10 +22,10 @@ function M.setup(_, opts)
           key
         )
       )
+      ---@diagnostic disable-next-line: no-unknown
       opts[key] = nil
     end
   end
-  format_opts = opts.format
   require("conform").setup(opts)
 end
 
@@ -53,7 +46,6 @@ return {
       },
     },
     init = function()
-      vim.o.formatexpr = "v:lua.require'conform'.formatexpr()"
       -- Install the conform formatter on VeryLazy
       require("lazyvim.util").on_very_lazy(function()
         require("lazyvim.util").format.register({
@@ -61,10 +53,14 @@ return {
           priority = 100,
           primary = true,
           format = function(buf)
-            require("conform").format(Util.merge(format_opts, { bufnr = buf }))
+            local plugin = require("lazy.core.config").plugins["conform.nvim"]
+            local Plugin = require("lazy.core.plugin")
+            local opts = Plugin.values(plugin, "opts", false)
+            require("conform").format(Util.merge(opts.format, { bufnr = buf }))
           end,
           sources = function(buf)
             local ret = require("conform").list_formatters(buf)
+            ---@param v conform.FormatterInfo
             return vim.tbl_map(function(v)
               return v.name
             end, ret)
@@ -82,19 +78,22 @@ return {
         }, { title = "LazyVim" })
       end
       ---@class ConformOpts
-      return {
+      local opts = {
         -- LazyVim will use these options when formatting with the conform.nvim formatter
         format = {
-          timeout_ms = 1000,
+          timeout_ms = 3000,
+          async = false, -- not recommended to change
+          quiet = false, -- not recommended to change
         },
+        ---@type table<string, conform.FormatterUnit[]>
         formatters_by_ft = {
           lua = { "stylua" },
           fish = { "fish_indent" },
           sh = { "shfmt" },
         },
-        -- LazyVim will merge the options you set here with builtin formatters.
+        -- The options you set here will be merged with the builtin formatters.
         -- You can also define any custom formatters here.
-        ---@type table<string,table>
+        ---@type table<string, conform.FormatterConfigOverride|fun(bufnr: integer): nil|conform.FormatterConfigOverride>
         formatters = {
           injected = { options = { ignore_errors = true } },
           -- # Example of using dprint only when a dprint.json file is present
@@ -106,10 +105,11 @@ return {
           --
           -- # Example of using shfmt with extra args
           -- shfmt = {
-          --   extra_args = { "-i", "2", "-ci" },
+          --   prepend_args = { "-i", "2", "-ci" },
           -- },
         },
       }
+      return opts
     end,
     config = M.setup,
   },