+++ lua/lazyvim/plugins/treesitter.lua
@@ -109,7 +109,7 @@ return {
     "nvim-treesitter/nvim-treesitter-context",
     event = "LazyFile",
     enabled = true,
-    opts = { mode = "cursor" },
+    opts = { mode = "cursor", max_lines = 3 },
     keys = {
       {
         "<leader>ut",
@@ -131,7 +131,7 @@ return {
   -- Automatically add closing tags for HTML and JSX
   {
     "windwp/nvim-ts-autotag",
-    event = "InsertEnter",
+    event = "LazyFile",
     opts = {},
   },
 }